Windows SMB servers do not support search pattern
Windows SMB servers require the pattern parameter to listing a path to
NOT be the specific path.  The shortcut previously used to get the
information about a specific path does not work with them, so list all
the contents of the containing directory, find the path in the list and
return it.